Topic controller (#25)
* added topic controller get by id

* added topic controller get by id

* fixed exception

* added topic controller - start topic

* archive topic

* topic controller - getActiveTopic

* fixed PR remarks

* test refactoring

+++ src/main/java/com/github/rodionovsasha/commentservice/controllers/TopicController.java
@@ -23,7 +23,7 @@ public class TopicController {
     }
 
     @GetMapping("/active/{id}")
-    public Topic getActiveTopic(@PathVariable long id) {
+    public Topic getActiveTopic(@PathVariable int id) {
         return service.getActiveTopic(id);
     }
